+++ RELEASES.md
@@ -23,6 +23,7 @@ Upcoming
 - Add attributes skipOnEmptyInput, ifParam and ifValue to FixedForwardPipe
 - Add value '!false' (equals 'true') to attribute active
 - Add PasswordHashPipe
+- Add check on queue existence in GetTibcoQueues and SendTibcoMessage (currently used in IJA_TiBeT2)